Make a IF argument to avoid error output
Checks if file exists. If it does, then run the script and delete the file. If not, do nothing.
Without this check there would be an error every time check-init.sh is run if the file is already deleted.

+++ vagrant/oc8ce/check-init.sh
@@ -66,8 +66,14 @@ GATEWAY=$(ip r | grep default | sed -e 's@.* dev \(\S\S*\) .*@\1@')
 ADDRESS=$(ip r | grep ${GATEWAY:-src} | grep src | head -n 1 | cut -d' ' -f12)
 
 ## set trusted domain on first boot, then delete the script.
-bash /var/scripts/set-trusted-domain.sh
-rm /var/scripts/set-trusted-domain.sh
+FILE="/var/scripts/set-trusted-domain.sh"
+if [ -f $FILE ];
+then
+        echo "Trusted domain is already set"
+else
+        bash /var/scripts/set-trusted-domain.sh
+        rm /var/scripts/set-trusted-domain.sh
+fi
 
 # set secure permissions
 bash /var/scripts/secure-permissions.sh
